# api/openapi/core.yaml
openapi: 3.0.3
info:
  title: Monival M&E API (Core)
  version: 0.1.0
  description: OpenAPI skeleton for core services (Auth, Tenants, Users, Projects, Indicators, Activities, Reporting).
servers:
  - url: /api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes: {}
...
paths:
...
  /reporting/periods:
    get:
      summary: List reporting periods
      security:
        - BearerAuth: []
    post:
      summary: Create reporting period
      security:
        - BearerAuth: []
...